broker: refactor the pkg structure
