+++ cmd/broker/core/testutils_test.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	"io"
@@ -8,21 +8,10 @@ import (
 
 // InsertFakeProcess inserts a fake Process into the broker for testing.
 // It does not start any OS-level process; it merely populates the broker map.
-func InsertFakeProcess(b *Broker, id string, stdin io.WriteCloser, stdout io.ReadCloser, status ProcessStatus) {
-	p := &Process{
-		info: &ProcessInfo{
-			ID:     id,
-			Status: status,
-		},
-		stdin:  stdin,
-		stdout: stdout,
-		done:   make(chan struct{}),
-	}
-
-	// Ensure the processes map is safe to write
-	b.mu.Lock()
-	b.processes[id] = p
-	b.mu.Unlock()
+func InsertFakeProcess(b *Broker, id string, stdin io.WriteCloser, stdout io.ReadCloser, status ProcessStatus) *Process {
+	p := NewTestProcess(id, status, stdin, stdout)
+	b.InsertProcessForTest(p)
+	return p
 }
 
 // ClosePipe closes both ends of an os.Pipe created by os.Pipe().
