+++ cmd/broker/spawn_adapter.go
@@ -1,5 +1,9 @@
 package main
 
+import (
+	"github.com/cyw0ng95/v2e/pkg/broker"
+)
+
 // SpawnAdapter delegates to the existing in-process Broker spawn methods.
 // This adapter is intentionally simple and non-invasive.
 type SpawnAdapter struct {
@@ -8,11 +12,11 @@ type SpawnAdapter struct {
 
 func NewSpawnAdapter(b *Broker) *SpawnAdapter { return &SpawnAdapter{b: b} }
 
-func toResult(info *ProcessInfo) *SpawnResult {
+func toResult(info *ProcessInfo) *broker.SpawnResult {
 	if info == nil {
 		return nil
 	}
-	return &SpawnResult{
+	return &broker.SpawnResult{
 		ID:       info.ID,
 		PID:      info.PID,
 		Command:  info.Command,
@@ -22,22 +26,25 @@ func toResult(info *ProcessInfo) *SpawnResult {
 	}
 }
 
-func (s *SpawnAdapter) Spawn(id, command string, args ...string) (*SpawnResult, error) {
+// Ensure SpawnAdapter implements broker.Spawner
+var _ broker.Spawner = (*SpawnAdapter)(nil)
+
+func (s *SpawnAdapter) Spawn(id, command string, args ...string) (*broker.SpawnResult, error) {
 	info, err := s.b.Spawn(id, command, args...)
 	return toResult(info), err
 }
 
-func (s *SpawnAdapter) SpawnRPC(id, command string, args ...string) (*SpawnResult, error) {
+func (s *SpawnAdapter) SpawnRPC(id, command string, args ...string) (*broker.SpawnResult, error) {
 	info, err := s.b.SpawnRPC(id, command, args...)
 	return toResult(info), err
 }
 
-func (s *SpawnAdapter) SpawnWithRestart(id, command string, maxRestarts int, args ...string) (*SpawnResult, error) {
+func (s *SpawnAdapter) SpawnWithRestart(id, command string, maxRestarts int, args ...string) (*broker.SpawnResult, error) {
 	info, err := s.b.SpawnWithRestart(id, command, maxRestarts, args...)
 	return toResult(info), err
 }
 
-func (s *SpawnAdapter) SpawnRPCWithRestart(id, command string, maxRestarts int, args ...string) (*SpawnResult, error) {
+func (s *SpawnAdapter) SpawnRPCWithRestart(id, command string, maxRestarts int, args ...string) (*broker.SpawnResult, error) {
 	info, err := s.b.SpawnRPCWithRestart(id, command, maxRestarts, args...)
 	return toResult(info), err
 }
